HealthBarView
HealthBarView done.
Rosie & Jeppe

+++ src/main/java/springboot/model/players/Player.java
@@ -109,7 +109,7 @@ public class Player {
         this.lives = lives + effect;
     }
 
-    public double getLives() {
+    public int getLives() {
         return lives;
     }
